# src/main/java/lu/dekabank/custodyconnect/controller/EventController.java
package lu.dekabank.custodyconnect.controller;

import java.util.ArrayList;
import java.util.List;
import java.util.Map;

import jakarta.inject.Inject;
import jakarta.validation.Valid;
import jakarta.ws.rs.Consumes;
import jakarta.ws.rs.DefaultValue;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.QueryParam;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;
import lu.dekabank.custodyconnect.model.Event;
import lu.dekabank.custodyconnect.model.dto.EventDto;
import lu.dekabank.custodyconnect.service.ReconciliationService;

@Path("/events")
@Consumes(MediaType.APPLICATION_JSON)
@Produces(MediaType.APPLICATION_JSON)
public class EventController {

  @Inject ReconciliationService reconciliationService;

  @POST
  public Response createEvent(@Valid EventDto dto) {
    try {
      Event event = reconciliationService.ingest(dto);
      return Response.accepted()
          .entity(Map.of("id", event.getId(), "message", "Event accepted for processing"))
          .build();
    } catch (Exception e) {
      return Response.serverError()
          .entity(Map.of("error", "Failed to process event", "message", e.getMessage()))
          .build();
    }
  }

  @GET
  public Response getEvents(
      @QueryParam("page") @DefaultValue("0") int page,
      @QueryParam("size") @DefaultValue("10") int size,
      @QueryParam("q") String query,
      @QueryParam("status") String status) {

    try {
      // Return basic response for now - simplify to avoid service dependencies
      List<Event> events = new ArrayList<>();
      long total = 0;

      Map<String, Object> response = Map.of(
        "items", events,
        "total", total,
        "page", page,
        "size", size,
        "totalPages", 0
      );

      return Response.ok(response).build();

    } catch (Exception e) {
      return Response.serverError()
          .entity(Map.of("error", "Failed to retrieve events", "message", e.getMessage()))
          .build();
    }
  }

  @GET
  @Path("/count")
  public Response getEventCount(
      @QueryParam("status") String status) {

    try {
      // Return basic count for now
      long total = 0;

      return Response.ok(Map.of("total", total)).build();

    } catch (Exception e) {
      return Response.serverError()
          .entity(Map.of("error", "Failed to count events", "message", e.getMessage()))
          .build();
    }
  }
}